% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_supplement.R
\name{install_supplement}
\alias{install_supplement}
\title{Install compressed remote data supplement locally}
\usage{
install_supplement(data_directory)
}
\arguments{
\item{data_directory}{Full path to the directory you wish to install the example data to.
Must be write-enabled for the user.}
}
\description{
Download and unpack example data supplement from Zenodo that matches the current installed
distribution.
}
