% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{log_config}
\alias{log_config}
\title{Mirror the configuration file to the log file}
\usage{
log_config(config, config_yml)
}
\arguments{
\item{config}{Configuration object from YAML read_yaml}

\item{config_yml}{Full path with file name to the configuration YAML file}
}
\description{
Mirror the configuration file to the log file
}
