% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{Run the citys2m model}
\usage{
model(config_yml, target_country_ids = NULL)
}
\arguments{
\item{config_yml}{Full path with file name to the configuration YAML file}

\item{target_country_ids}{Optional.  List of country ids to evaluate.}
}
\description{
Run the citys2m model
}
