% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{urban_neighbor}
\alias{urban_neighbor}
\title{Calculate the neighborhood using a window}
\usage{
urban_neighbor(urbanMap, winSize)
}
\arguments{
\item{urbanMap}{binary urban map (0-nonubran; 1-urban)}

\item{winSize}{window size}
}
\value{
a raster of neighborhood value indicating urban probability
}
\description{
Calculate the neighborhood using a window
}
