% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_model.R
\name{prepare_model}
\alias{prepare_model}
\title{Prepare the ubran data for the citys2m model}
\usage{
prepare_model(config_yml = NULL, config_obj = NULL,
  write_output = TRUE)
}
\arguments{
\item{config_yml}{Full path with file name to the configuration YAML file}

\item{config_obj}{Configuration object from \code{read_yaml}}

\item{write_output}{Optional.  If true, the output will be written to file.}
}
\description{
Prepare the ubran data for the citys2m model
}
